backend/requirements.txt
alembic==1.8.1
anyio==3.6.1
asttokens==2.0.8
Authlib==1.0.1
backcall==0.2.0
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
executing==1.0.0
fastapi==0.81.0
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
jedi==0.18.1
Mako==1.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
Pygments==2.13.0
python-dotenv==0.20.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
stack-data==0.3.0
starkbank-ecdsa==2.0.3
starlette==0.19.1
traitlets==5.3.0
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
